renaming server and index #93 #94
